+++ cmake/modules/Element.cmake
@@ -20,6 +20,7 @@ target_sources(kwik_element
             modules/element/checkbox.cppm
             modules/element/textarea.cppm
             modules/element/dropdown.cppm
+            modules/element/slider.cppm
 
 
         BASE_DIRS ${CMAKE_CURRENT_SOURCE_DIR}/modules/layout
@@ -42,6 +43,7 @@ target_sources(kwik_element
         src/element/checkbox.cpp
         src/element/textarea.cpp
         src/element/dropdown.cpp
+        src/element/slider.cpp
 
         src/layout/flex_layout.cpp
         src/layout/grid_layout.cpp